# openapi/auto-gestao-api.yaml
openapi: 3.0.3
info:
  title: auto-gestao-api
  description: >
    Backend do domínio de condomínio (Condomínio, Endereço, Pessoa, Taxa Extra,
    Meio de Contato) do ecossistema Gestão Condial. Contrato contract-first:
    este arquivo é a fonte da verdade dos endpoints; as interfaces de
    controller e os modelos são gerados a partir dele pelo
    openapi-generator-maven-plugin.
  version: "1.0.0"
servers:
  - url: /
...
paths:
...
  /pessoa:
    get:
      tags: [Pessoa]
      operationId: findAll
      parameters:
        - name: idCondominio
          in: query
          description: Restringe a um condomínio.
          schema:
            type: integer
            format: int64
        - name: idUnidade
          in: query
          description: Restringe a uma unidade.
          schema:
            type: integer
            format: int64
        - name: semUnidade
          in: query
          description: Quando verdadeiro, traz só os moradores ainda não vinculados a uma unidade.
          schema:
            type: boolean
        - name: busca
          in: query
          description: Texto livre; casa com nome e CPF.
          schema:
            type: string
      summary: Lista todas as pessoas cadastradas.
      responses:
        "200":
          description: Lista de pessoas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pessoa"
    post:
      tags: [Pessoa]
      operationId: savePessoa
      summary: Cria uma nova pessoa.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Pessoa"
      responses:
        "201":
          description: Pessoa criada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pessoa"
        "400":
          description: Erro de validação dos parâmetros enviados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListaDeErros"
  /pessoa/{idPessoa}:
    get:
      tags: [Pessoa]
      operationId: findPessoaId
      summary: Busca uma pessoa pelo id.
      parameters:
        - name: idPessoa
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Pessoa encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pessoa"
    put:
      tags: [Pessoa]
      operationId: updatePessoa
      summary: Atualiza um morador.
      description: >
        Exige a permissão `MORADOR_ESCREVER`. Para desvincular o morador da unidade, envie
        `unidade` nulo.
      parameters:
        - name: idPessoa
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Pessoa"
      responses:
        "200":
          description: Morador atualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pessoa"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    delete:
      tags: [Pessoa]
      operationId: deletePessoa
      summary: Exclui um morador.
      description: Exige a permissão `MORADOR_ESCREVER`. Os meios de contato dele vão junto.
      parameters:
        - name: idPessoa
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Morador excluído.
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /pessoa/findEnvioPorEmailPorCondominio/{id}:
    get:
      tags: [Pessoa]
      operationId: findEnvioPorEmailPorCondominio
      summary: Lista as pessoas de um condomínio que recebem taxa por e-mail.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Lista de pessoas com envio de taxa por e-mail habilitado.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pessoa"
  /pessoa/listPessoasCondominio/{idCondominio}:
    get:
      tags: [Pessoa]
      operationId: findPessoasCondominio
      summary: Lista as pessoas de um condomínio.
      parameters:
        - name: idCondominio
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Lista de pessoas do condomínio.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pessoa"
  /pessoa/meioContatoPessoa:
    post:
      tags: [Pessoa]
      operationId: saveMeioContato
      summary: Cria um meio de contato para uma pessoa.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeioContato"
      responses:
        "201":
          description: Meio de contato criado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeioContato"
        "400":
          description: Erro de validação dos parâmetros enviados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListaDeErros"
  /unidade:
    get:
      tags: [Unidade]
      operationId: listUnidades
      summary: Lista as unidades.
      description: Exige a permissão `UNIDADE_LER`.
      parameters:
        - name: idCondominio
          in: query
          description: Restringe a um condomínio.
          schema:
            type: integer
            format: int64
        - name: ativa
          in: query
          schema:
            type: boolean
        - name: busca
          in: query
          description: Texto livre; casa com bloco e número.
          schema:
            type: string
      responses:
        "200":
          description: Unidades encontradas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unidade"
    post:
      tags: [Unidade]
      operationId: createUnidade
      summary: Cadastra uma unidade.
      description: >
        Exige a permissão `UNIDADE_ESCREVER`. A unidade nasce sem morador — vincular alguém é um
        passo separado, feito pelo cadastro do morador.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Unidade"
      responses:
        "201":
          description: Unidade criada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unidade"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "409":
          description: Já existe uma unidade com esse bloco e número no condomínio.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListaDeErros"
  /unidade/{idUnidade}:
    get:
      tags: [Unidade]
      operationId: findUnidade
      summary: Busca uma unidade pelo id.
      description: Exige a permissão `UNIDADE_LER`.
      parameters:
        - name: idUnidade
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Unidade encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unidade"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    put:
      tags: [Unidade]
      operationId: updateUnidade
      summary: Atualiza uma unidade.
      description: Exige a permissão `UNIDADE_ESCREVER`.
      parameters:
        - name: idUnidade
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Unidade"
      responses:
        "200":
          description: Unidade atualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unidade"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          description: Já existe outra unidade com esse bloco e número no condomínio.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListaDeErros"
    delete:
      tags: [Unidade]
      operationId: deleteUnidade
      summary: Exclui uma unidade.
      description: >
        Exige a permissão `UNIDADE_ESCREVER`. Recusa com 409 se houver morador vinculado —
        desvincule antes, para que a remoção seja uma decisão explícita e não um efeito colateral.
      parameters:
        - name: idUnidade
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Unidade excluída.
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/EmUso"
components:
  responses:
    DadosInvalidos:
      description: Dados inválidos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ListaDeErros"
    NaoEncontrado:
      description: Recurso não encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ListaDeErros"
    EmUso:
      description: O recurso ainda está vinculado a outros registros.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ListaDeErros"
  schemas:
    Unidade:
      type: object
      description: >
        Unidade autônoma de um condomínio. Existe independentemente de haver morador: o vínculo
        com uma Pessoa é opcional nos dois sentidos.

        Sem `required` de propósito: este mesmo schema aparece aninhado em `Pessoa.unidade`,
        onde serve só para referenciar a unidade pelo `id`. Exigir `numero` e `condominio` aqui
        rejeitaria `{"unidade": {"id": 7}}` na hora de vincular um morador. A obrigatoriedade
        vale na criação da unidade e é aplicada no `UnidadeDto`.
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        condominio:
          $ref: "#/components/schemas/Condominio"
        bloco:
          type: string
          maxLength: 30
          description: Opcional — condomínios de bloco único não usam.
        numero:
          type: string
          maxLength: 20
          description: Identificação da unidade. Texto, não número — "101", "A-3" e "Loja 2" valem.
        andar:
          type: integer
          format: int32
        fracaoIdeal:
          type: number
          format: double
          description: Participação da unidade no rateio das despesas.
        observacao:
          type: string
          maxLength: 300
        ativa:
          type: boolean
          default: true
        quantidadeDeMoradores:
          type: integer
          format: int32
          readOnly: true
          description: Quantos moradores estão vinculados. Zero é condição para poder excluir.
    Condominio:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        ddd:
          type: integer
          format: int32
        numeroTelefone:
          type: integer
          format: int64
        endereco:
          $ref: "#/components/schemas/Endereco"
        valorJuros:
          type: number
          format: double
        valorMulta:
          type: number
          format: double
        possuiTaxaExtra:
          type: boolean
          default: false
        valorTaxaCondominial:
          type: number
          format: double
        valorTaxasExtras:
          type: number
          format: double
          readOnly: true
          description: Somatório das taxas extras vigentes, calculado pelo servidor.
        valorTotalPorUnidade:
          type: number
          format: double
          readOnly: true
          description: Valor final por unidade (taxa condominial + taxas extras), calculado pelo servidor.
    Endereco:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        cep:
          type: integer
          format: int64
        endereco:
          type: string
        complemento:
          type: string
    Pessoa:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        nomeCompleto:
          type: string
        unidade:
          allOf:
            - $ref: "#/components/schemas/Unidade"
          description: >
            Unidade onde a pessoa mora. Nulo é válido — um morador pode ser cadastrado antes de
            se saber onde vai morar. Para vincular, basta enviar o objeto com o `id`.
        envioTaxaEmail:
          type: boolean
        envioImpresso:
          type: boolean
        isSindico:
          type: boolean
        condominio:
          $ref: "#/components/schemas/Condominio"
        endereco:
          $ref: "#/components/schemas/Endereco"
        cpf:
          type: string
          description: CPF da pessoa (validado no servidor).
    MeioContato:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        pessoa:
          $ref: "#/components/schemas/Pessoa"
        tipoMeioContato:
          type: string
          enum:
            - EMAIL
            - TELEF
        identificacao:
          type: string
...
    ErroDeParametro:
      type: object
      properties:
        parametro:
          type: string
        mensagem:
          type: string
    ListaDeErros:
      type: object
      properties:
        erros:
          type: array
          items:
            $ref: "#/components/schemas/ErroDeParametro"
        quantidadeDeErros:
          type: integer
          format: int32
          readOnly: true